+++ src/fast-querystring.ts
@@ -39,7 +39,8 @@ import fastDecode from 'fast-decode-uri-component'
 
 const plusRegex = /\+/g
 
-export function parseQuery(input: string) {
+// Parse query without array
+export function parseQueryFromURL(input: string) {
 	const result = <Record<string, string>>{}
 
 	if (typeof input !== 'string') return result
@@ -124,6 +125,111 @@ export function parseQuery(input: string) {
 
 				result[key] = value
 			}
+		}
+	}
+
+	return result
+}
+
+/**
+ * @callback parse
+ * @param {string} input
+ */
+export const parseQuery = (input: string) => {
+	const result: Record<string, string[]> = {}
+
+	if (typeof input !== 'string') return result
+
+	const inputLength = input.length
+	let key = ''
+	let value = ''
+	let startingIndex = -1
+	let equalityIndex = -1
+	let shouldDecodeKey = false
+	let shouldDecodeValue = false
+	let keyHasPlus = false
+	let valueHasPlus = false
+	let hasBothKeyValuePair = false
+	let c = 0
+
+	// Have a boundary of input.length + 1 to access last pair inside the loop.
+	for (let i = 0; i < inputLength + 1; i++) {
+		if (i !== inputLength) c = input.charCodeAt(i)
+		else c = 38
+
+		// Handle '&' and end of line to pass the current values to result
+		switch (c) {
+			case 38: {
+				hasBothKeyValuePair = equalityIndex > startingIndex
+
+				// Optimization: Reuse equality index to store the end of key
+				if (!hasBothKeyValuePair) equalityIndex = i
+
+				key = input.slice(startingIndex + 1, equalityIndex)
+
+				// Add key/value pair only if the range size is greater than 1; a.k.a. contains at least "="
+				if (hasBothKeyValuePair || key.length > 0) {
+					// Optimization: Replace '+' with space
+					if (keyHasPlus) key = key.replace(plusRegex, ' ')
+
+					// Optimization: Do not decode if it's not necessary.
+					if (shouldDecodeKey) key = fastDecode(key) || key
+
+					if (hasBothKeyValuePair) {
+						value = input.slice(equalityIndex + 1, i)
+
+						if (valueHasPlus) value = value.replace(plusRegex, ' ')
+
+						if (shouldDecodeValue)
+							value = fastDecode(value) || value
+					}
+
+					const currentValue = result[key]
+
+					if (currentValue === undefined)
+						// @ts-ignore - As current value is undefined, we can safely assign it
+						result[key] = value
+					else {
+						// @ts-ignore - Optimization: value.pop is faster than Array.isArray(value)
+						if (currentValue.pop) currentValue.push(value)
+						// @ts-ignore - As current value is a string, convert it to an array
+						else result[key] = [currentValue, value]
+					}
+				}
+
+				// Reset reading key value pairs
+				value = ''
+				startingIndex = i
+				equalityIndex = i
+				shouldDecodeKey = false
+				shouldDecodeValue = false
+				keyHasPlus = false
+				valueHasPlus = false
+
+				break
+			}
+
+			// Check '='
+			case 61:
+				if (equalityIndex <= startingIndex) equalityIndex = i
+				// If '=' character occurs again, we should decode the input.
+				else shouldDecodeValue = true
+
+				break
+
+			// Check '+', and remember to replace it with empty space.
+			case 43:
+				if (equalityIndex > startingIndex) valueHasPlus = true
+				else keyHasPlus = true
+
+				break
+
+			// Check '%' character for encoding
+			case 37:
+				if (equalityIndex > startingIndex) shouldDecodeValue = true
+				else shouldDecodeKey = true
+
+				break
 		}
 	}